docs: replace sailboat icon in README
The previous icon was from the Noun Project [0], and although I had
purchased a license, I'm not sure if it's able to be used as a
logo for a project like this?

The new icon is a modified version of a CC0 icon from pixabay[1],
so there shouldn't be any licensing problems.

[0]: https://thenounproject.com/icon/528354/
[1]: https://pixabay.com/en/sailing-sailing-boat-catboat-99229/